third_party/blink/web_tests/platform/win11-arm64/http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-in-http-header-expected.txt

CONSOLE MESSAGE: Script executed in iframe.
CONSOLE MESSAGE: FAIL: Iframe was not in a unique origin: I am a secret
